backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
cohere==4.3.1
qdrant-client>=1.7.0
asyncpg==0.29.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
python-multipart==0.0.6
cryptography==41.0.8